# docs/swagger.yaml
basePath: /
definitions:
...
  model.CommentCreateReq:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  model.CommentDeleteRes:
    properties:
      comment_id:
        type: string
    type: object
  model.CommentResponse:
    properties:
      comment_id:
        type: string
      created_at:
        type: string
      message:
        type: string
      photo_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.CommentUpdateReq:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  model.CommentUpdateRes:
    properties:
      comment_id:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      code:
        type: integer
      errors: {}
      status:
        type: string
    type: object
...
  model.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a MyGram API.
  title: MyGram API
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: get all comment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CommentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: get all comment
      tags:
      - Comment
  /comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: delete comment
      parameters:
      - description: comment_id
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CommentDeleteRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: delete comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: get one comment
      parameters:
      - description: comment_id
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: get one comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: update comment
      parameters:
      - description: comment_id
        in: path
        name: comment_id
        required: true
        type: string
      - description: request is required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CommentUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CommentUpdateRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: update comment
      tags:
      - Comment
  /comments/{photo_id}:
    post:
      consumes:
      - application/json
      description: create comment for a particular user
      parameters:
      - description: photo_id
        in: path
        name: photo_id
        required: true
        type: string
      - description: request is required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CommentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: create comment
      tags:
      - Comment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"